Change qreals to floats
+ Fix default axes to sensible
+ Fix some rounding errors in surface creation

Task-number: QTRD-2622

+++ examples/bars/graphmodifier.h
@@ -61,13 +61,13 @@ signals:
 
 private:
     Q3DBars *m_graph;
-    qreal m_xRotation;
-    qreal m_yRotation;
+    float m_xRotation;
+    float m_yRotation;
     int m_fontSize;
     int m_segments;
     int m_subSegments;
-    qreal m_minval;
-    qreal m_maxval;
+    float m_minval;
+    float m_maxval;
     QStringList m_months;
     QStringList m_years;
     Q3DValueAxis *m_temperatureAxis;
